requirements.txt
asgiref==3.7.2
Django==4.2.2
django-taggit==4.0.0
importlib-metadata==6.7.0
Markdown==3.4.3
sitemap==20191121
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
zipp==3.15.0
